scalefs, create: Fix accumulation of extra link counts on the parent directory
When the 'create' operation of a directory is flushed, a new directory inode is
created, and its contents are initialized with a ".." link pointing to its parent.
At some later point in time, when the parent is fsynced, the link operation that
adds the newly created directory to the parent's directory entries will be flushed.

The net effect of both these operations should be to increment the parent
directory's link count by 1 (due to the addition of 1 sub-directory). However,
we used to increment the parent's link count at both these stages, leading to an
extra link count on the parent. This was also true in the case of directory-renames;
changing the ".." pointer in the sub-directory being moved would increment the
link count on the destination parent directory incorrectly.

Fix this by never changing any link counts when adding the ".." link to a newly
created directory.

+++ kernel/fs.cc
@@ -1011,12 +1011,15 @@ dirlink(sref<inode> dp, const char *name, u32 inum, bool inc_link,
 
   dp->dir_offset += sizeof(struct dirent);
 
-  sref<inode> ip = iget(1, inum);
-  if (ip)
-    ip->link();
+  // If adding the ".." link in a directory, don't change *any* link counts.
+  if (strncmp(name, "..", DIRSIZ) != 0) {
+    sref<inode> ip = iget(1, inum);
+    if (ip)
+      ip->link();
 
-  if (inc_link)
-    dp->link();
+    if (inc_link)
+      dp->link();
+  }
 
   dir_flush_entry(dp, name, trans);
   return 0;
@@ -1039,12 +1042,15 @@ dirunlink(sref<inode> dp, const char *name, u32 inum, bool dec_link,
   if (!dp->dir->insert(strbuf<DIRSIZ>(name), de_info))
     return -1;
 
-  sref<inode> ip = iget(1, inum);
-  if (ip)
-    ip->unlink();
+  // If removing the ".." link in a directory, don't change *any* link counts.
+  if (strncmp(name, "..", DIRSIZ) != 0) {
+    sref<inode> ip = iget(1, inum);
+    if (ip)
+      ip->unlink();
 
-  if (dec_link)
-    dp->unlink();
+    if (dec_link)
+      dp->unlink();
+  }
 
   dir_flush_entry(dp, name, trans);
   dp->dir->remove(strbuf<DIRSIZ>(name));
